inc/dec: fix R behavior

+++ src/core/cpu.c
@@ -581,17 +581,17 @@ void tx_cpu_op_pop(tx_CPU* cpu, tx_Parameters* params) {
 void tx_cpu_op_inc(tx_CPU* cpu, tx_Parameters* params) {
     AR_SIMPLE_OP_1("inc", 1 +);
     R_SIZES(
-        ((result.i == INT32_MIN) << 1) | (result.u == 0),
+        ((result.i == INT8_MIN) << 1) | (result.u == 0),
         ((result.i == INT16_MIN) << 1) | (result.u == 0),
-        ((result.i == INT8_MIN) << 1) | (result.u == 0)
+        ((result.i == INT32_MIN) << 1) | (result.u == 0)
     );
 }
 void tx_cpu_op_dec(tx_CPU* cpu, tx_Parameters* params) {
     AR_SIMPLE_OP_1("dec", -1 +);
     R_SIZES(
-        ((result.i == INT32_MAX) << 1) | (result.u == UINT32_MAX),
+        ((result.i == INT8_MAX) << 1) | (result.u == UINT8_MAX),
         ((result.i == INT16_MAX) << 1) | (result.u == UINT16_MAX),
-        ((result.i == INT8_MAX) << 1) | (result.u == UINT8_MAX)
+        ((result.i == INT32_MAX) << 1) | (result.u == UINT32_MAX)
     );
 }
 void tx_cpu_op_add(tx_CPU* cpu, tx_Parameters* params) { AR_OVF_OP(add) }
